Move solidity code and tooling to contracts/solidity workspace
