+++ apps/website/src/composables/contracts.ts
@@ -1,0 +1,10 @@
+// import Sample from '@casimir/solidity/build/artifacts/contracts/src/Sample.sol/Sample.json'
+// import { ethers } from 'ethers'
+
+export default function useContracts() {
+
+
+    return {
+
+    }
+}
