+++ contracts/solidity/hardhat.config.ts
@@ -1,0 +1,30 @@
+import { task } from 'hardhat/config'
+import '@typechain/hardhat'
+import '@nomiclabs/hardhat-waffle'
+import '@nomiclabs/hardhat-ethers'
+import { HardhatUserConfig } from 'hardhat/config'
+
+// https://hardhat.org/guides/create-task.html
+task('accounts', 'Prints the list of accounts', async (taskArgs, hre) => {
+  const accounts = await hre.ethers.getSigners()
+
+  for (const account of accounts) {
+    console.log(account.address)
+  }
+})
+
+// Go to https://hardhat.org/config/ to learn more
+const config: HardhatUserConfig = {
+  solidity: '0.8.4',
+  paths: {
+    tests: './test',
+    sources: './src',
+    artifacts: './build/artifacts',
+    cache: './build/cache'
+  },
+  typechain: {
+    outDir: './build/artifacts/types'
+  }
+}
+
+export default config
