+++ contracts/solidity/package.json
@@ -1,0 +1,26 @@
+{
+    "name": "@casimir/solidity",
+    "scripts": {
+        "deploy:sample": "npx hardhat run deploy/sample.ts",
+        "test:contracts": "npx hardhat test",
+        "task:accounts": "npx hardhat accounts",
+        "task:compile": "npx hardhat compile",
+        "task:clean": "npx hardhat clean",
+        "node:local": "npx hardhat node"
+    },
+    "devDependencies": {
+        "@nomiclabs/hardhat-ethers": "^2.0.6",
+        "@nomiclabs/hardhat-waffle": "^2.0.3",
+        "@typechain/ethers-v5": "^10.1.0",
+        "@typechain/hardhat": "^6.1.2",
+        "@types/chai": "^4.3.1",
+        "@types/mocha": "^9.1.1",
+        "@types/node": "^17.0.45",
+        "chai": "^4.3.6",
+        "ethereum-waffle": "^3.4.4",
+        "ethers": "^5.6.9",
+        "hardhat": "^2.9.9",
+        "ts-node": "^10.8.2",
+        "typechain": "^8.1.0"
+    }
+}
