+++ contracts/solidity/src/Sample.sol
@@ -1,0 +1,22 @@
+//SPDX-License-Identifier: Unlicense
+pragma solidity ^0.8.4;
+
+import "hardhat/console.sol";
+
+contract Sample {
+    string private greeting;
+
+    constructor(string memory _greeting) {
+        console.log("Deploying a Sample with greeting:", _greeting);
+        greeting = _greeting;
+    }
+
+    function greet() public view returns (string memory) {
+        return greeting;
+    }
+
+    function setGreeting(string memory _greeting) public {
+        console.log("Changing greeting from '%s' to '%s'", greeting, _greeting);
+        greeting = _greeting;
+    }
+}
