+++ contracts/solidity/tsconfig.json
@@ -1,0 +1,19 @@
+{
+    "compilerOptions": {
+        "target": "es2018",
+        "module": "commonjs",
+        "strict": true,
+        "esModuleInterop": true,
+        "outDir": "dist"
+    },
+    "include": [
+        "./deploy/*.ts",
+        "./test/*.ts"
+    ],
+    "typeRoots": [
+        "node_modules/@types"
+    ],
+    "files": [
+        "hardhat.config.ts"
+    ]
+}
