+++ package.json
@@ -5,6 +5,7 @@
   "workspaces": [
     "apps/*",
     "common/*",
+    "contracts/*",
     "infrastructure/*",
     "services/*"
   ],
@@ -16,12 +17,12 @@
     "test": "echo \"Error: no test specified\" && exit 1",
     "test:push": "scripts/actions/test -w push",
     "release": "release-it",
-    "deploy:sample": "npx hardhat run scripts/hardhat/sample.ts",
-    "test:contracts": "npx hardhat test",
-    "task:accounts": "npx hardhat accounts",
-    "task:compile": "npx hardhat compile",
-    "task:clean": "npx hardhat clean",
-    "node:local": "npx hardhat node"
+    "deploy:sample": "npm run deploy:sample --workspace @casimir/solidity",
+    "test:contracts": "npm run test:contracts --workspace @casimir/solidity",
+    "task:accounts": "npm run task:accounts --workspace @casimir/solidity",
+    "task:compile": "npm run task:compile --workspace @casimir/solidity",
+    "task:clean": "npm run task:clean --workspace @casimir/solidity",
+    "node:local": "npm run node:local --workspace @casimir/solidity"
   },
   "repository": {
     "type": "git",
@@ -35,28 +36,15 @@
   },
   "homepage": "https://github.com/consensusnetworks/casimir#readme",
   "devDependencies": {
-    "@nomiclabs/hardhat-ethers": "^2.0.6",
-    "@nomiclabs/hardhat-waffle": "^2.0.3",
-    "@typechain/ethers-v5": "^10.1.0",
-    "@typechain/hardhat": "^6.1.2",
-    "@types/chai": "^4.3.1",
-    "@types/mocha": "^9.1.1",
-    "@types/node": "^17.0.45",
     "@typescript-eslint/eslint-plugin": "^5.30.5",
     "@typescript-eslint/parser": "^5.30.5",
     "@vue/eslint-config-typescript": "^10.0.0",
-    "chai": "^4.3.6",
     "eslint": "^8.19.0",
     "eslint-plugin-vue": "^8.7.1",
-    "ethereum-waffle": "^3.4.4",
-    "ethers": "^5.6.9",
-    "hardhat": "^2.9.9",
     "html-to-text": "^8.2.0",
     "lint-staged": "^13.0.0",
     "mjml": "^4.12.0",
-    "release-it": "^15.0.0",
-    "ts-node": "^10.8.2",
-    "typechain": "^8.1.0"
+    "release-it": "^15.0.0"
   },
   "lint-staged": {
     "*.{ts,vue}": "npm run lint"
