--- tsconfig.json
@@ -1,19 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "es2018",
-        "module": "commonjs",
-        "strict": true,
-        "esModuleInterop": true,
-        "outDir": "dist"
-    },
-    "include": [
-        "./scripts/**/*.ts",
-        "./contracts/**/*.ts"
-    ],
-    "typeRoots": [
-        "node_modules/@types"
-    ],
-    "files": [
-        "./hardhat.config.ts"
-    ]
-}
